Time: 4 ms (50.00%), Space: 42.9 MB (32.81%) - LeetHub

+++ 781-rabbits-in-forest/781-rabbits-in-forest.java
@@ -11,9 +11,9 @@ class Solution {
         
         for(Map.Entry<Integer, Integer> e: hm.entrySet()) {
             int groupSize = e.getKey()+1;
-            int freq = e.getValue();
+            double freq = e.getValue();
             
-            int numOfGroups = (int)Math.ceil((double)freq/groupSize);
+            int numOfGroups = (int)Math.ceil(freq/groupSize);
             int rabbitsInCurrGroup = numOfGroups * groupSize;
             
             res += rabbitsInCurrGroup;
